Update Set_Path_Selection_Policies
This script Connects to vCenter, requires the name of the ESXi Hosts, requires the ESXi Hosts are in Maintenance Mode, Identifies the disks that are not set to Round Robin and are not vendor  'HP' (ie not local) disks.  It then applies the change and lists the ESXi Hosts it has been applied to.